[Gold I] Title: 최솟값과 최댓값, Time: 160 ms, Memory: 5536 KB -BaekjoonHub

+++ 백준/Gold/2357. 최솟값과 최댓값/최솟값과 최댓값.cc
@@ -4,7 +4,7 @@
 
 using namespace std;
 
-long long n, a[100000], seg1[400000], seg2[400000];
+int n, a[100000], seg1[400000], seg2[400000];
 
 int construct1(int start, int end, int idx)
 {
